Fix DatasmithExporter using 100% of one logical core.
#rb none

[CL 15581614 by benoit deschenes in Main branch]

+++ Engine/Source/Developer/Datasmith/DatasmithExporter/Private/DatasmithExporterManager.cpp
@@ -184,7 +184,7 @@ uint32 FDatasmithGameThread::Run()
 		}
 
 		// Update this if we can get refresh rate of the user monitor(s)
-		constexpr double FrameTime = 1 / 60;
+		constexpr double FrameTime = (double)1 / 60;
 		double TimeSinceLastUpdate =  FPlatformTime::Seconds() - LastSlateUpdateTime;
 		while (TimeSinceLastUpdate < FrameTime && bKeepRunning)
 		{
